requirements.txt
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.5
docker==7.0.0
Flask==2.2.5
Flask-JWT-Extended==4.4.3
Flask-SQLAlchemy==2.5.1
greenlet==3.0.3
grpcio==1.51.1
gunicorn==22.0.0
idna==3.7
importlib_metadata==7.1.0
itsdangerous==2.2.0
Jinja2==3.1.3
joblib==1.4.2
MarkupSafe==2.1.5
msgpack==1.0.8
numpy==1.24.4
nvflare==2.4.0
packaging==24.0
pandas==2.0.3
patsy==0.5.6
protobuf==3.20.3
psutil==5.9.8
pycparser==2.22
pyhocon==0.3.60
PyJWT==2.8.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
scikit-learn==1.3.2
scipy==1.10.1
six==1.16.0
SQLAlchemy==1.4.31
statsmodels==0.14.1
threadpoolctl==3.5.0
tzdata==2024.1
urllib3==2.2.1
websockets==12.0
Werkzeug==2.2.2
zipp==3.18.1
dominate==2.9.1
simplejson==3.19.3
